+++ R/flist.R
@@ -375,6 +375,119 @@ as.data.frame.fmt_lst <- function(x, row.names = NULL, optional = FALSE, ...) {
 }
 
 
+
+# Read and Write flist ----------------------------------------------------
+
+
+
+#' @title Write a formatting list to the file system
+#' @description The \code{write.flist} function writes a formatting list
+#' to the file system.  By default, the formatting list will be written to the 
+#' current working directory, using the variable name as the file name.  These
+#' defaults can be overridden using the appropriate parameters.  The catalog
+#' will be saved with a file extension of ".flist". 
+#' 
+#' Note that the formatting list is saved as an RDS file.  The ".flist" file 
+#' extension only serves to distinguish the format catalog from other RDS
+#' files.
+#' @param x The formatting list to write.
+#' @param dir_path The directory path to write the catalog to. Default is the 
+#' current working directory.
+#' @param file_name The name of the file to save the catalog as.  Default is
+#' the name of the variable that contains the formatting list.  The ".flist" file
+#' extension will be added automatically.
+#' @return The full path of the saved formatting list.
+#' @family flist
+#' @examples 
+#' # Create formatting list
+#' fl <- flist(f1 = "%5.1f",
+#'             f2 = "%6.2f",
+#'             type = "row")
+#'            
+#' # Get temp directory
+#' tmp <- tempdir()            
+#'            
+#' # Save formatting list to file system
+#' pth <- write.flist(fl, dir_path = tmp)
+#' 
+#' # Read from file system
+#' fr <- read.flist(pth)
+#' 
+#' # Create sample data
+#' dat <- c(12.3844, 292.28432)
+#' 
+#' # Use formats in the catalog
+#' fapply(dat, fr)
+#' # [1] " 12.4"  "292.28"
+#' 
+#' @export
+write.flist <- function(x, dir_path = getwd(), file_name = NULL) {
+  
+  if (is.null(file_name))
+    file_name <- deparse(substitute(x, env = environment()))
+  
+  pth <- file.path(dir_path, paste0(file_name, ".flist"))
+  
+  
+  if (file.exists(pth))
+    file.remove(pth)
+  
+  saveRDS(x, pth)
+  
+  
+  log_logr("Saved formatting list to '" %p% pth %p% "'")
+  
+  return(pth)
+}
+
+
+#' @title Read a formatting list from the file system
+#' @description The \code{read.flist} function reads a formatting list
+#' from the file system.  The function accepts a path to the formatting list,
+#' reads the list, and returns it.
+#' 
+#' Note that the formatting list is saved as an RDS file.  The ".flist" file 
+#' extension only serves to distinguish the formatting list from other RDS
+#' files.
+#' @param file_path The path to the formatting list.
+#' @return The formatting list as an R object.
+#' @family flist
+#' @examples 
+#' # Create formatting list
+#' fl <- flist(f1 = "%5.1f",
+#'             f2 = "%6.2f",
+#'             type = "row")
+#'            
+#' # Get temp directory
+#' tmp <- tempdir()            
+#'            
+#' # Save formatting list to file system
+#' pth <- write.flist(fl, dir_path = tmp)
+#' 
+#' # Read from file system
+#' fr <- read.flist(pth)
+#' 
+#' # Create sample data
+#' dat <- c(12.3844, 292.28432)
+#' 
+#' # Use formats in the catalog
+#' fapply(dat, fr)
+#' # [1] " 12.4"  "292.28"
+#' @export
+read.flist <- function(file_path) {
+  
+  ret <-  readRDS(file_path)
+  
+  log_logr("Read formatting list from '" %p% file_path %p% "'")
+  
+  if (log_output()) {
+    log_logr(ret)
+    print(ret)
+  }
+  return(ret)
+}
+
+
 #' @title Print a formatting list
 #' @param x The formatting list to print
 #' @param ... Follow-on parameters to the print function